:hammer: rename package from 'treinamentofiec' to 'trainingfiec'
