+++ core/src/main/java/com/adobe/aem/guides/wknd/core/models/trainingfiec/Banner.java
@@ -1,4 +1,4 @@
-package com.adobe.aem.guides.wknd.core.models.treinamentofiec;
+package com.adobe.aem.guides.wknd.core.models.trainingfiec;
 
 import org.apache.sling.api.resource.Resource;
 import org.apache.sling.models.annotations.DefaultInjectionStrategy;
