+++ core/src/main/java/com/adobe/aem/guides/wknd/core/models/trainingfiec/package-info.java
@@ -1,4 +1,4 @@
 @Version("2.0.0")
-package com.adobe.aem.guides.wknd.core.models.treinamentofiec;
+package com.adobe.aem.guides.wknd.core.models.trainingfiec;
 
 import org.osgi.annotation.versioning.Version;
