+++ ui.apps/src/main/content/jcr_root/apps/wknd/clientlibs/clientlib-fiec/.content.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <jcr:root xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
     jcr:primaryType="cq:ClientLibraryFolder"
-    categories="[wknd.fiec]"
+    categories="[wknd.training-fiec]"
     cssProcessor="[default:none,min:none]"
     jsProcessor="[default:none,min:none]"
     allowProxy="{Boolean}true"/>
