+++ ui.apps/src/main/content/jcr_root/apps/wknd/components/banner/banner.html
@@ -1,10 +1,10 @@
 <sly data-sly-use.clientlib="/libs/granite/sightly/templates/clientlib.html">
-	<sly data-sly-call="${clientlib.css @ categories='wknd.fiec'}" />
+	<sly data-sly-call="${clientlib.css @ categories='wknd.training-fiec'}" />
 </sly>
 
 <div
 	class="banner-container"
-	data-sly-use.banner="com.adobe.aem.guides.wknd.core.models.treinamentofiec.Banner"
+	data-sly-use.banner="com.adobe.aem.guides.wknd.core.models.trainingfiec.Banner"
 >
 	<div class="banner-content">
 		<div class="banner-text-container">
